src/main/steps/ca-da-respondent/support-you-need-during-case/documents-support/content.test.ts
import languageAssertions from '../../../../../test/unit/utils/languageAssertions';
import { FormContent, FormFields, FormOptions } from '../../../../app/form/Form';
import { CommonContent } from '../../../common/common.content';

import { generateContent } from './content';

const en = {
  section: 'Reasonable adjustments',
  title: 'I need documents in an alternative format',
  courtCommunication:
    'Think about all communications with the court, as well as what you might need at a hearing. Consider remote and in-person hearings, in case your preferred hearing type is not possible.',
  optionHint: 'Select all that apply to you',
  summaryText: 'Contacts for help',
  docsColour: 'Documents in a specified colour',
  docsColourDetails: 'Describe what you need',
  docsReadFormat: 'Documents in an easy read format',
  docsReadFormatHint: 'information written in simple language with pictures',
  brailleDocs: 'Braille documents',
  largePrintDocs: 'Documents in large print',
  largePrintDocsDetails: 'Describe what you need',
  audioTranslation: 'Audio translation of documents',
  docsReadOut: 'Documents read out to me',
  emailInfo: 'Information emailed to me',
  other: 'Other',
  otherDetails: 'Describe what you need',
  noSupport: 'I do not need any of this support at this time',
  continue: 'Save and continue',
  errors: {
    respondentDocsSupport: {
      required: 'Please select an answer',
    },
    respondentDocsDetails: {
      required: 'Please provide the docs details',
    },
    respondentLargePrintDetails: {
      required: 'Please provide the large print details',
    },
    respondentOtherDetails: {
      required: 'Please provide the other details',
    },
  },
};

const cy: typeof en = {
  section: 'Reasonable adjustments',
  title: 'I need documents in an alternative format',
  courtCommunication:
    'Think about all communications with the court, as well as what you might need at a hearing. Consider remote and in-person hearings, in case your preferred hearing type is not possible.',
  optionHint: 'Select all that apply to you',
  summaryText: 'Contacts for help',
  docsColour: 'Documents in a specified colour',
  docsColourDetails: 'Describe what you need',
  docsReadFormat: 'Documents in an easy read format',
  docsReadFormatHint: 'information written in simple language with pictures',
  brailleDocs: 'Braille documents',
  largePrintDocs: 'Documents in large print',
  largePrintDocsDetails: 'Describe what you need',
  audioTranslation: 'Audio translation of documents',
  docsReadOut: 'Documents read out to me',
  emailInfo: 'Information emailed to me',
  other: 'Other',
  otherDetails: 'Describe what you need',
  noSupport: 'I do not need any of this support at this time',
  continue: 'Save and continue',
  errors: {
    respondentDocsSupport: {
      required: 'Please select an answer',
    },
    respondentDocsDetails: {
      required: 'Please provide the docs details',
    },
    respondentLargePrintDetails: {
      required: 'Please provide the large print details',
    },
    respondentOtherDetails: {
      required: 'Please provide the other details',
    },
  },
};

jest.mock('../../../../app/form/validation');
/* eslint-disable @typescript-eslint/ban-types */
describe('citizen-home content', () => {
  const commonContent = { language: 'en' } as CommonContent;
  let generatedContent;
  let form;
  let fields;
  beforeEach(() => {
    generatedContent = generateContent(commonContent);
    form = generatedContent.form as FormContent;
    fields = form.fields as FormFields;
  });

  test('should return correct english content', () => {
    expect(generatedContent.title).toEqual('I need documents in an alternative format');
    expect(generatedContent.section).toEqual('Reasonable adjustments');
    expect(generatedContent.courtCommunication).toEqual(
      'Think about all communications with the court, as well as what you might need at a hearing. Consider remote and in-person hearings, in case your preferred hearing type is not possible.'
    );
    expect(generatedContent.optionHint).toEqual('Select all that apply to you');
    expect(generatedContent.summaryText).toEqual('Contacts for help');
    expect(generatedContent.docsColour).toEqual('Documents in a specified colour');
    expect(generatedContent.docsReadFormat).toEqual('Documents in an easy read format');
    expect(generatedContent.brailleDocs).toEqual('Braille documents');
    expect(generatedContent.largePrintDocs).toEqual('Documents in large print');
    expect(generatedContent.audioTranslation).toEqual('Audio translation of documents');
    expect(generatedContent.docsReadOut).toEqual('Documents read out to me');
    expect(generatedContent.emailInfo).toEqual('Information emailed to me');
    expect(generatedContent.other).toEqual('Other');
    expect(generatedContent.noSupport).toEqual('I do not need any of this support at this time');
  });

  // eslint-disable-next-line jest/expect-expect
  test('should return correct english content Data', () => {
    languageAssertions('en', en, () => generateContent(commonContent));
  });

  // eslint-disable-next-line jest/expect-expect
  test('should return correct welsh content', () => {
    languageAssertions('cy', cy, () => generateContent({ ...commonContent, language: 'cy' }));
  });

  test('should contain docsSupport field', () => {
    const docsSupportField = fields.respondentDocsSupport as FormOptions;
    expect(docsSupportField.type).toBe('checkboxes');
    expect((docsSupportField.section as Function)(generatedContent)).toBe(en.section);
  });

  test('should contain Save and continue button', () => {
    expect((form.submit?.text as Function)(generatedContent)).toBe('Save and continue');
  });
});

/* eslint-enable @typescript-eslint/ban-types */
